Refactor(SE-62): Consolidate property listing components into one new component and import it to the page
